Add usegalaxy.eu badge
cuteSV is now available on the European Galaxy server. The European Galaxy server is part of the German Network of Bioinformatics Infrastructure (de.NBI), ELIXIR, and the European Science Cloud (EOSC-life). We hope that this badge will increase the cuteSV userbase in the community by pointing user to a combination of a graphical user-interface and compute resources.

Thanks for all the effort that has been put into cuteSV and for making it open source.